Implement foundations for profile content items
Sets up module to create a new "Profile" content type that has a
"ProfilePart" that is responsible for linking the profile content item
to an Orchard user. The initial migration will create a profile for
all existing users. Currently there is no way to add a handler for when
a user is created, however that will be available when a new stable v
version of Orchard Core is released.

By default a "Profile" can't be created via the admin dashboard.