Improve admin validation
We originally had Permissions.ViewContent to check if the user is an admin but actually this permission is for anonymous users as well so it has changed to EditContent now so the Orchard Admin editors only can view the restricted contents.

I’ve also added redirect to the root of the website if there is no redirect URL has been specified in the settings.