Add Profile Group Field
Add a new feature ProfileGroupField which enables a Field populated by
content items from types which implement the ProfileGroupPart. This part
can be added to content types or parts to allow them to be associated with
groups. The part has been implemented with a view to having a multiple
select option in future but currently only allows single select through a
drop down. Multi select could be added using a different Editor mode.

Add Model, Driver, ViewModels and Views for the new part.

Update the README to describe the new Feature.

Add the None display mode which doesn't render anything on the front end.