Add Required to ProfileGroupField
Add a Required setting to ProfileGroupField. When checked the field will
show 'Please select' in place of 'None' and the driver will check that at
least 1 group has been selected and display an error if it hasn't.

Fix issues with display mode not being stored by adding another display
mode hyphenated type and using that for 'Standard' this shouldn't be
required but it wasn't working otherwise.

Rename None display mode to Zero.