Add ProfileGroupOwnershipPart access restriction
Add access restriction features to the ProfileGroupOwnershipPart allow the
editor to specify if the Part should restrict users without the correct
profile group from accessing the content. This can be set at the whole
type level or at the item level by the editor.

Add a field to ProfileGroupOwnershipSettings for RestrictAccess which is
an enum with options None, Type and Item.

Add an editor view for the part which will have the restrict access
checkbox when the mode is set to Item.

Add IOwnershipAuthorizationService inspired by IAccessAuthorizationService
in SubscriptionAccessGrouping. This service will be responsible for
checking if a given user should have access to content based on their
profile group membership.