Add helpful GroupOwnership features
Add more helpful features to GroupOwnership needed when using it in a
project.

Add ContentItemExtensions currently with an extension method for getting
a list of COntentItemIds of owning groups from a ContentItem.

Update OwnershipAuthorizationService to return false if the user is null.

Add GroupOwnershipIndex for fast querying of owned content items and
update migration to add this table.