Fix database lock issue when removing profile from group
Noticed when working on NCC that we were getting lots of database lock
issues when trying to assign profiles to exhibitors. Looking through
previous commits it showed that Mobin had a similar issue which he
resolved with the following commit:

https://bitbucket.org/etchuk/etch.orchardcore.userprofiles/commits/70a040d6017f8526e1efa35f49698110fb67cd2b

Unfortunately it doesn't seem to have solved the issue fully. Playing
with the code I seemed to fix the lock issue by adding a call to
`PublishAsync`.