Add an auto complete picker to GroupField
Add an auto completing picker field based on the content picker
implementation used in the profile picker from the Grouping feature to the
GroupField.

Refactor the front end scripts to output in directories per feature as
index.js.

Add a new script for profileGroupPicker which uses the existing
profilePicker but now passes in a search endpoint and a boolean for if the
picker should allow multiples.

Update profilePicker to accept the endpoint for autocomplete results
defaulting to the current ProfilePicker one and also accept a boolean for
if multiple selections should be allowed, defaulting to true. The usage in
the profile picker is therefore unchanged.

Add the controller to provide search results for the ProfileGroups to the
GroupField feature package. Also add the PickerResultProvider.